+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.2
+
+* added config (file only) for 
+  * laser x/y width
+  * border speed
+* fix "extra data path" dialog problem
+
+
 # v1.1
 
 * fix pad for rotated footprints